+++ product/src/main/java/store/_0982/product/application/dto/ProductUpdateCommand.java
@@ -4,7 +4,7 @@ import store._0982.product.domain.ProductCategory;
 
 public record ProductUpdateCommand(
         String name,
-        int price,
+        Long price,
         ProductCategory category,
         String description,
         int stock,
